refactored and finally got the shadow
