+++ src/components/blocks/contact.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Contact = () => {
+  return <div>contact</div>;
+};
