+++ src/components/blocks/index.ts
@@ -1,0 +1,4 @@
+export * from "./home";
+export * from "./about";
+export * from "./portfolio";
+export * from "./experience";
