+++ src/page.tsx
@@ -1,20 +1,13 @@
 import React from "react";
-import { Box, Container } from "@chakra-ui/react";
+import { Box } from "@chakra-ui/react";
 import { Header } from "./components/header";
 import Main from "./components/main";
 
 const Page = () => {
   return (
-    <Box bg="#272b23" pb="100px">
-      <Container maxW="1000px">
-        <Box>
-          <Header />
-        </Box>
-        <Box>
-          <Main />
-        </Box>
-        <Box>{/* Footer */}</Box>
-      </Container>
+    <Box bg="#282929" pb="100px">
+      <Header />
+      <Main />
     </Box>
   );
 };
